@@ -13,14 +13,18 @@
     "version": "2.8.1",
     "type": "project",
     "require": {
-        "php": "^7.3|^8.0",
-        "chrome-php/chrome": "^1.9",
+        "php": "^8.3",
+        "ext-curl": "*",
+        "ext-dom": "*",
+        "ext-pdo": "*",
+        "chrome-php/chrome": "^1.14",
         "doctrine/dbal": "^3.5",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.0.1",
         "intervention/image": "^2.7",
         "ksubileau/color-thief-php": "^2.0",
         "laravel/framework": "^8.75",
+        "laravel/horizon": "^5.22",
         "laravel/tinker": "^2.5",
         "league/flysystem-aws-s3-v3": "^1.0",
         "php-open-source-saver/jwt-auth": "^2.0",
