@@ -2,44 +2,42 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-
-use App\Models\Post;
+use App\Http\Requests\Post\PostDeleteRequest;
+use App\Http\Requests\Post\PostIndexRequest;
+use App\Http\Requests\Post\PostShowRequest;
+use App\Http\Requests\Post\PostStoreRequest;
+use App\Http\Requests\Post\PostUpdateRequest;
+use App\Http\Requests\Post\PostUrlInfoRequest;
 use App\Models\Collection;
+use App\Models\Post;
 use App\Models\User;
-use App\Models\Tag;
+use App\Repositories\Contracts\CollectionRepositoryInterface;
+use App\Repositories\Contracts\PostRepositoryInterface;
+use App\Repositories\Contracts\TagRepositoryInterface;
 use App\Services\PostService;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class PostController extends Controller
 {
-
-    private $service;
-
-    public function __construct()
+    public function __construct(
+        protected PostService $service,
+        protected PostRepositoryInterface $repository,
+        protected CollectionRepositoryInterface $collectionRepository,
+        protected TagRepositoryInterface $tagRepository,
+    )
     {
-        $this->service = new PostService();
+        //
     }
 
-    public function index(Request $request)
+    /**
+     * @param PostIndexRequest $request
+     * @return JsonResponse
+     */
+    public function index(PostIndexRequest $request): JsonResponse
     {
-
-        $this->validate($request, [
-            'collection_id'    => 'integer|nullable',
-            'is_uncategorized' => 'nullable',
-            // should support: 0, 1, true, false, "true", "false" because
-            // it should/could be used in a query string
-            'tag_id'           => 'integer|nullable',
-            'withTags'         => 'nullable',
-            'filter'           => 'string|nullable',
-            'is_archived'      => 'nullable',
-            // same as is_uncategorized
-            'after_id'         => 'integer|nullable',
-            'offset'           => 'integer|nullable',
-            'limit'            => 'integer|nullable',
-        ]);
-
         $auth_type = User::getAuthenticationType();
 
         if ($auth_type === User::UNAUTHORIZED_USER) {
@@ -47,116 +45,110 @@ class PostController extends Controller
         }
 
         $after_post = null;
-        if (isset($request->after_id)) {
-            if (!isset($request->collection_id) && isset($request->is_uncategorized)) {
+        if ($request->getAfterId()) {
+            if (!$request->getCollectionId() && !$request->getIsUncategorized()) {
                 return response()->json(
                     'collection_id or is_uncategorized is required',
                     Response::HTTP_BAD_REQUEST
                 );
             }
-            $after_post = Post::find($request->after_id);
+            $after_post = $this->repository->getById($request->getAfterId());
             if ($after_post == null) {
                 return response()->json('after_id does not exist', Response::HTTP_NOT_FOUND);
             }
+
             $this->authorize('view', $after_post);
             $collection_id = Collection::getCollectionId(
-                $request->collection_id,
-                $request->is_uncategorized
+                $request->getCollectionId(),
+                $request->getIsUncategorized()
             );
+
             if ($after_post->collection_id !== $collection_id) {
                 return response()->json('Wrong collection', Response::HTTP_BAD_REQUEST);
             }
         }
 
-        if (isset($request->tag_id)) {
-            if (!Tag::find($request->tag_id)->exists()) {
+        if ($request->getTagId()) {
+            if (!$this->tagRepository->getById($request->getTagId())) {
                 return response()->json('Tag does not exist', Response::HTTP_BAD_REQUEST);
             }
         }
 
         $posts = $this->service->all(
-            intval($request->collection_id),
-            $this->service->boolValue($request->is_uncategorized),
-            intval($request->tag_id),
-            $this->service->boolValue($request->withTags),
-            strval($request->filter),
+            Auth::id(),
+            $request->getCollectionId(),
+            $request->getIsUncategorized(),
+            $request->getTagId(),
+            $request->getWithTags(),
+            $request->getFilter(),
             $auth_type,
-            $this->service->boolValue($request->is_archived),
+            $request->getIsArchived(),
             $after_post,
-            intval($request->offset),
-            intval($request->limit),
+            $request->getOffset(),
+            $request->getLimit(),
         );
 
-        return response()->json(['data' => $posts], Response::HTTP_OK);
+        return response()->json(['data' => $posts]);
     }
 
-    public function show(int $id, Request $request)
+    /**
+     * @param int $id
+     * @param PostShowRequest $request
+     * @return JsonResponse
+     */
+    public function show(int $id, PostShowRequest $request): JsonResponse
     {
-
-        $this->validate($request, [
-            'withTags' => 'nullable',
-        ]);
-        $withTags = $this->service->boolValue($request->withTags);
-
-        $post = null;
-
-        if ($withTags) {
-            $post = Post::with('tags:id,name')->find($id);
-        } else {
-            $post = Post::find($id);
-        }
+        $post = $this->repository->getById($id, $request->getWithTags());
 
         if ($post === null) {
             return response()->json('Post does not exist', Response::HTTP_NOT_FOUND);
         }
+
         $this->authorize('view', $post);
 
         return response()->json(['data' => $post], Response::HTTP_OK);
     }
 
-    public function store(Request $request)
+    /**
+     * @param PostStoreRequest $request
+     * @return JsonResponse
+     */
+    public function store(PostStoreRequest $request): JsonResponse
     {
-        $this->validate($request, [
-            'title'         => 'string|nullable',
-            'content'       => 'required|string',
-            'collection_id' => 'integer|nullable',
-            'description'   => 'string|nullable',
-            'tags'          => 'array|nullable',
-            'tags.*'        => 'integer',
-        ]);
+        if ($request->getCollectionId()) {
+            $collection = $this->collectionRepository->getById($request->getCollectionId());
 
-        if (isset($request->collection_id)) {
-            $collection = Collection::findOrFail($request->collection_id);
-            if (Auth::user()->id !== $collection->user_id) {
+            if (!$collection) {
+                throw new ModelNotFoundException(
+                    'Collection does not exist',
+                );
+            }
+
+            if (Auth::id() !== $collection->user_id) {
                 return response()->json('Not authorized', Response::HTTP_FORBIDDEN);
             }
         }
 
         $post = $this->service->store(
-            $request->title,
-            $request->content,
-            $request->collection_id,
-            $request->description,
-            $request->tags,
-            Auth::user()->id
+            Auth::id(),
+            $request->getPostContent(),
+            $request->getTitle(),
+            $request->getCollectionId(),
+            $request->getDescription(),
+            $request->getTags(),
         );
 
         return response()->json(['data' => $post], Response::HTTP_CREATED);
     }
 
-    public function update(Request $request, $id)
+    /**
+     * @param PostUpdateRequest $request
+     * @param int $id
+     * @return JsonResponse
+     */
+    public function update(PostUpdateRequest $request, int $id)
     {
-
-        $validatedData = $this->validate($request, [
-            'title'            => 'string|nullable',
-            'content'          => 'string|nullable',
-            'collection_id'    => 'integer|nullable',
-            'is_uncategorized' => 'boolean|nullable',
-            'tags'             => 'array|nullable',
-            'tags.*'           => 'integer|required_with:tags',
-            'order'            => 'integer|nullable',
-            'is_archived'      => 'boolean|nullable'
-        ]);
+        $validatedData = $request->validated();
 
         $post = Post::withTrashed()->find($id);
         if (!$post) {
@@ -165,25 +157,29 @@ class PostController extends Controller
 
         $this->authorize('update', $post);
 
-        $request->is_uncategorized = boolval($request->is_uncategorized);
-
-        if (empty($request->collection_id) && $request->is_uncategorized === false) {
+        if ($request->getCollectionId() && $request->getIsUncategorized() === false) {
             // request contains no knowledge about a collection
             $validatedData['collection_id'] = $post->collection_id;
         } else {
             $validatedData['collection_id'] = Collection::getCollectionId(
-                $request->collection_id,
-                $request->is_uncategorized
+                $request->getCollectionId(),
+                $request->getIsUncategorized()
             );
         }
 
         if (!empty($validatedData['collection_id'])) {
-            Collection::findOrFail($validatedData['collection_id']);
+            $collection = $this->collectionRepository->getById($validatedData['collection_id']);
+            if (!$collection) {
+                throw new ModelNotFoundException(
+                    'Collection does not exist',
+                );
+            }
         }
 
         if (isset($validatedData['content'])) {
             $validatedData['content'] = $this->service->sanitize($validatedData['content']);
-            $info = $this->service->computePostData($request->title, $validatedData['content']);
+            $info = $this->service->computePostData($validatedData['content'], $request->getTitle());
+
             if ($validatedData['content'] === $post->content) {
                 foreach (['title', 'description', 'image_path'] as $attr) {
                     if (!empty($post->{$attr})) {
@@ -197,48 +193,61 @@ class PostController extends Controller
         }
 
         $newValues = array_merge($validatedData, $info);
-        $newValues['user_id'] = Auth::user()->id;
+        $newValues['user_id'] = Auth::id();
 
         if ($post->collection_id !== $validatedData['collection_id']) {
             // post wants to have a different collection than before
             // compute order in new collection
-            $newValues['order'] = Post::where('collection_id', $validatedData['collection_id'])
-                ->max('order') + 1;
+            $newValues['order'] = Post::query()
+                    ->where('collection_id', $validatedData['collection_id'])
+                    ->max('order') + 1;
+
             // reorder old collection
-            Post::where('collection_id', $post->collection_id)
-                ->where('order', '>', $post->order)->decrement('order');
+            Post::query()
+                ->where('collection_id', $post->collection_id)
+                ->where('order', '>', $post->order)
+                ->decrement('order');
+
         } else if (isset($validatedData['order'])) {
             // post wants to be positioned somewhere else
             // staying in the same collection as before
             $newOrder = $validatedData['order'];
 
             // check authenticity of order
-            if (!Post::where('collection_id', $post->collection_id)->where('order', $newOrder)->exists()) {
-                $newOrder = Post::where('collection_id', $post->collection_id)->max('order');
+            if (!Post::query()
+                ->where('collection_id', $post->collection_id)
+                ->where('order', $newOrder)->exists()
+            ) {
+                $newOrder = Post::query()
+                    ->where('collection_id', $post->collection_id)
+                    ->max('order');
                 $newValues['order'] = $newOrder;
             }
 
             $oldOrder = $post->order;
             if ($newOrder !== $oldOrder) {
                 if ($newOrder > $oldOrder) {
-                    Post::where('collection_id', $post->collection_id)
-                        ->whereBetween('order', [$oldOrder + 1, $newOrder])->decrement('order');
+                    Post::query()
+                        ->where('collection_id', $post->collection_id)
+                        ->whereBetween('order', [$oldOrder + 1, $newOrder])
+                        ->decrement('order');
                 } else {
-                    Post::where('collection_id', $post->collection_id)
-                        ->whereBetween('order', [$newOrder, $oldOrder - 1])->increment('order');
+                    Post::query()
+                        ->where('collection_id', $post->collection_id)
+                        ->whereBetween('order', [$newOrder, $oldOrder - 1])
+                        ->increment('order');
                 }
             }
         }
 
         $post->update($newValues);
 
-        if (isset($request->is_archived)) {
-            $request->is_archived = filter_var($request->is_archived, FILTER_VALIDATE_BOOLEAN);
+        if ($request->getIsArchived()) {
             $is_currently_archived = $post->trashed();
-            if ($request->is_archived === true && $is_currently_archived === false) {
+            if ($request->getIsArchived() === true && $is_currently_archived === false) {
                 $this->service->delete($post);
                 $post = Post::withTrashed()->find($post->id);
-            } else if ($request->is_archived === false && $is_currently_archived === true) {
+            } else if ($request->getIsArchived() === false && $is_currently_archived === true) {
                 $post = $this->service->restore($post);
             }
         }
@@ -251,19 +260,23 @@ class PostController extends Controller
         if (isset($newValues['tags'])) {
             $this->service->saveTags($post->id, $newValues['tags']);
         }
-        $post->tags = $post->tags()->get();
 
-        return response()->json(['data' => $post], Response::HTTP_OK);
+        return response()->json(['data' => $post->refresh()], Response::HTTP_OK);
     }
 
-    public function destroy(int $id)
+    /**
+     * @param int $id
+     * @param PostDeleteRequest $request
+     * @return JsonResponse
+     */
+    public function destroy(int $id, PostDeleteRequest $request): JsonResponse
     {
-
-        $post = Post::find($id);
+        $post = $this->repository->getById($id);
 
         if (!$post) {
             return response()->json('Post not found.', Response::HTTP_NOT_FOUND);
         }
+
         $this->authorize('delete', $post);
 
         $this->service->delete($post);
@@ -271,12 +284,12 @@ class PostController extends Controller
         return response()->json('', Response::HTTP_NO_CONTENT);
     }
 
-    public function getUrlInfo(Request $request)
+    /**
+     * @param PostUrlInfoRequest $request
+     * @return JsonResponse
+     */
+    public function getUrlInfo(PostUrlInfoRequest $request): JsonResponse
     {
-        $this->validate($request, [
-            'url' => 'url'
-        ]);
-
-        return response()->json($this->service->getInfo($request->url));
+        return response()->json($this->service->getInfo($request->getUrl()));
     }
 }
