@@ -27,9 +27,9 @@ class InstallCommand extends Command
     /**
      * Execute the console command.
      *
-     * @return mixed
+     * @return void
      */
-    public function handle()
+    public function handle(): void
     {
 
         $createOnlyUser = $this->option('only-user');
