@@ -2,14 +2,37 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
+use App\Scopes\TagScope;
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
+/**
+ * @property int $id
+ * @property int $user_id
+ * @property string $name
+ *
+ * @property Carbon|null $created_at
+ * @property Carbon|null $updated_at
+ * @property Carbon|null $deleted_at
+ *
+ * @property-read Collection|Post[] $posts
+ */
 class Tag extends Model
 {
-    public $timestamps = false;
     use SoftDeletes, HasFactory;
+
+    /**
+     * @return void
+     */
+    public static function boot(): void
+    {
+        parent::boot();
+
+        static::addGlobalScope(new TagScope);
+    }
 
     /**
      * The attributes that should be cast to native types.
@@ -17,7 +40,11 @@ class Tag extends Model
      * @var array
      */
     protected $casts = [
-        'user_id' => 'integer', // because of SQLite
+        'user_id' => 'integer',
+        'name' => 'string',
+        'created_at' => 'datetime',
+        'updated_at' => 'datetime',
+        'deleted_at' => 'datetime'
     ];
 
     /**
@@ -26,7 +53,8 @@ class Tag extends Model
      * @var array
      */
     protected $fillable = [
-        'name', 'user_id'
+        'user_id',
+        'name',
     ];
 
     /**
@@ -35,13 +63,17 @@ class Tag extends Model
      * @var array
      */
     protected $hidden = [
-        'user_id', 'deleted_at', 'pivot'
+        'user_id',
+        'pivot',
+        'created_at',
+        'updated_at',
+        'deleted_at',
     ];
 
     /**
-     * Posts that belong to the tag.
+     * @return BelongsToMany
      */
-    public function posts()
+    public function posts(): BelongsToMany
     {
         return $this->belongsToMany(Post::class);
     }
