@@ -11,23 +11,24 @@ class CreatePostsTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('posts', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->id();
+
+            $table->foreignId('user_id')->constrained()->cascadeOnUpdate()->cascadeOnDelete();
+            $table->foreignId('collection_id')->nullable()->constrained()->cascadeOnUpdate()->cascadeOnDelete();
+
             $table->text('content');
             $table->tinyInteger('type');
             $table->text('url')->nullable();
             $table->string('title')->nullable();
             $table->text('description')->nullable();
-            $table->string('color', 7)->nullable();
+            $table->string('color', 40)->nullable();
             $table->string('image_path')->nullable();
             $table->string('base_url')->nullable();
-            $table->unsignedInteger('collection_id')->nullable();
-            $table->foreign('collection_id')->references('id')->on('collections');
-            $table->unsignedInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('users');
             $table->unsignedSmallInteger('order');
+
             $table->timestamps();
             $table->softDeletes();
         });
@@ -38,7 +39,7 @@ class CreatePostsTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('posts');
     }
