@@ -11,18 +11,17 @@ class CreateSharesTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('shares', function (Blueprint $table) {
-            $table->increments('id');
+            $table->id();
             $table->text('token');
-            $table->unsignedInteger('collection_id')->nullable();
-            $table->foreign('collection_id')->references('id')->on('collections');
-            $table->unsignedBigInteger('post_id')->nullable();
-            $table->foreign('post_id')->references('id')->on('posts');
+
+            $table->foreignId('collection_id')->constrained()->cascadeOnUpdate()->cascadeOnDelete();
+            $table->foreignId('post_id')->constrained()->cascadeOnUpdate()->cascadeOnDelete();
+            $table->foreignId('created_by')->constrained('users')->cascadeOnUpdate()->cascadeOnDelete();
+
             $table->unsignedTinyInteger('permission')->default(4);
-            $table->foreign('created_by')->references('id')->on('users');
-            $table->unsignedInteger('created_by');
             $table->boolean('is_active');
         });
     }
@@ -32,7 +31,7 @@ class CreateSharesTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('shares');
     }
